+++ src/writer/fail_on_skipped.rs
@@ -60,18 +60,19 @@ where
     Wr: for<'val> ArbitraryWriter<'val, W, String>,
 {
     async fn handle_event(&mut self, ev: parser::Result<event::Cucumber<W>>) {
-        use event::{Cucumber, Feature, Rule, Scenario, Step};
+        use event::{
+            Cucumber, Feature, Rule, Scenario, Step, StepError::Panic,
+        };
 
-        let map_failed =
-            |f: Arc<_>, r: Option<Arc<_>>, sc: Arc<_>, st: Arc<_>| {
-                let event = if (self.should_fail)(&f, r.as_deref(), &sc) {
-                    Step::Failed(None, None, Arc::new("not allowed to skip"))
-                } else {
-                    Step::Skipped
-                };
-
-                Ok(Cucumber::scenario(f, r, sc, Scenario::Step(st, event)))
+        let map_failed = |f: Arc<_>, r: Option<Arc<_>>, sc: Arc<_>, st: _| {
+            let event = if (self.should_fail)(&f, r.as_deref(), &sc) {
+                Step::Failed(None, None, Panic(Arc::new("not allowed to skip")))
+            } else {
+                Step::Skipped
             };
+
+            Ok(Cucumber::scenario(f, r, sc, Scenario::Step(st, event)))
+        };
 
         let ev = match ev {
             Ok(Cucumber::Feature(